Updates to latest stuff and gets dashboard running
This gets the dashboard running via HDMI output to monitor form RPI3
The touchscreen still shows no life at all.
I am not sure if it is my touchscreen or I am missing something else to
make that work with this. I am really starting to think it is the screen
itself that is the issue. I have never got anything to display not even
the logging stuff on bootup.
The dashboard has toolshed and init-gadget and is setup to connect over ETH0
for ssh via nerves.local.

@tapickell