Dummy Light Simple added
In the end I learned about using the cache and hash for images
and that the path at compile time is different than the path at runtime.
So in the init make sure to get the runtime path to load the image to the rect.

In the end I opted for just text that is orange and only shows
when the on_state is true.

This still needs to handle codes and displaying them as well.

@tapickell