Got pattern matching on at processor
This adds a hex module with the hx/1 macro
This is for having a clean way to define these
letters as hex values for pattern matching.
This may not be the best way to do this.
It works for now and should be revisited and refactored later.

Also alpha's the at command functions for easier finding
and adds stub functions on serial and responder respond/1

@tapickell