Trying to graph communication flow
also adds more notes
need to figure out more about the commuication
between the rpi and the ecu
I have done this before with Ruby using a gem and
with arduino and python back in the day.
Need to refresh memory on how these things talk
Is it Elm instructions to the ECU via the adapter?